app/requirements.txt
fastapi==0.115.11
python-multipart==0.0.20
uvicorn==0.34.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
Pillow==11.1.0
numpy==2.2.3
easyocr==1.7.2
mistralai==1.5.1
pydantic==2.10.6
ollama==0.6.0